made it integrate with the other because of Checking teams fault

+++ FlyPlusJourneyContentEnricher/Models/DTO/FlightBookingDTO.cs
@@ -15,6 +15,6 @@ namespace FlyPlusJourneyContentEnricher.Models.DTO
         public string FlightOrigin { get; set; }
         public string FlightDestination { get; set; }
         public int PassengersAvailableTotal { get; set; }
-        public double BaggageWeightAvailableTotal { get; set; }
+        public int BaggageWeightAvailableTotal { get; set; }
     }
 }
